Moved CPU Emulation Engine options to the Advanced tab
